common컨트롤러 관련 로직수정
